02-modularization/pyproject.toml
[project]
name = "fastapi-modularization"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "faker (>=37.11.0,<38.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=5.0.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
